pyproject.toml
[project]
name = "bfe-rueckliefertarif"
version = "0.1.0"
description = "Home Assistant integration for Swiss PV Rückliefertarif (Art. 15 EnFV)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "bfe-rueckliefertarif contributors" }]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-homeassistant-custom-component>=0.13",
    "aioresponses>=0.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "RUF"]
ignore = ["E501"]
